Add support to apply labels when upstream issue is closed
Make it possible to configure a list of labels to be applied on Jira
cards, when the corresponding upstream issue is closed.

This can come handy, when one doesn't want to automatically close
downstream issues when upstream issues are closed, but still want's to
have an easy way to filter for such issues in Jira.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>