Use jira format and gfm when converting desc
While syncing an issue to Jira, if the `github_markdown` option is
listed in `issue_updates`, and the upstream issue is from GitHub, then
use the proper GitHub-Flavored Markdown parser (`gfm`) and covert it to
the `jira` format instead of `plain` text.

This improves how things like links and code blocks are represented in
the created Jira issue.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>